+++ cli/cmds/deployment.js
@@ -4,7 +4,7 @@ exports.command = 'deployment <create|upgrade|destroy|config>';
 exports.desc = 'manage the wharfie service deployment';
 exports.builder = function (yargs) {
   return yargs
-    .commandDir('deployment_cmds')
+    .commandDir('./deployment_cmds')
     .demandCommand()
     .showHelpOnFail(true);
 };
