+++ cli/cmds/deployment_cmds/config.js
@@ -1,104 +1,51 @@
 'use strict';
 
-const child_process = require('child_process');
 const inquirer = require('inquirer');
-const CloudFormation = require('../../../lambdas/lib/cloudformation');
-const STS = require('../../../lambdas/lib/sts');
 const { displayFailure, displayInfo, displaySuccess } = require('../../output');
-
-const deployment_template = require('../../../cloudformation/deployment/wharfie.template.js');
-
-const cloudformation = new CloudFormation();
-const sts = new STS();
+const WharfieDeployment = require('../../../lambdas/lib/actor/wharfie-deployment');
 
 const config = async (development) => {
-  const template = deployment_template;
-  const stackName = process.env.WHARFIE_DEPLOYMENT_NAME;
   const answers = await new Promise((resolve, reject) => {
     inquirer
-      .prompt(
-        Object.keys(template.Parameters).reduce((acc, key) => {
-          if (['Version', 'IsDevelopment'].includes(key)) {
-            return acc;
-          }
-          const p = template.Parameters[key];
-          let type = 'input';
-          if (p.Type === 'Number') {
-            type = 'number';
-          } else if (p.AllowedValues) {
-            type = 'list';
-          }
-          if (key === 'ArtifactBucket' && development) {
-            delete p.Default;
-          } else if (key === 'ArtifactBucket' && !development) {
-            return acc;
-          }
-          if (key === 'GitSha') {
-            if (development) {
-              p.Default = child_process
-                .execSync('git rev-parse HEAD')
-                .toString()
-                .trim();
-            } else {
-              return acc;
-            }
-          }
-          return acc.concat({
-            name: key,
-            message: p.Description,
-            default: p.Default,
-            type,
-            choices: p.AllowedValues,
-          });
-        }, [])
-      )
+      .prompt([
+        {
+          type: 'number',
+          name: 'globalQueryConcurrency',
+          message:
+            'enter the maximum number of concurrent queries this deployment can run',
+          default: 10,
+        },
+        {
+          type: 'number',
+          name: 'resourceQueryConcurrency',
+          message:
+            'enter the maximum number of concurrent queries any resource in this deployment can run',
+          default: 10,
+        },
+        {
+          type: 'number',
+          name: 'maxQueriesPerAction',
+          message:
+            'enter the maximum number of total queries any action can request',
+          default: 10000,
+        },
+        {
+          type: 'list',
+          name: 'loggingLevel',
+          message: 'enter the logging level for the deployment',
+          choices: ['debug', 'info', 'warn', 'error'],
+          default: 'info',
+        },
+      ])
       .then(resolve)
       .catch(reject);
   });
-  const { Stacks } = await cloudformation.describeStacks({
-    StackName: stackName,
-  });
-  const version = Stacks[0].Parameters.find(
-    (p) => p.ParameterKey === 'Version'
-  ).ParameterValue;
   displayInfo(`updating wharfie deployment configuration...`);
-  const { Account } = await sts.getCallerIdentity();
-  await cloudformation.updateStack({
-    StackName: stackName,
-    Tags: [],
-    Parameters: [
-      ...Object.keys(answers).map((key) => {
-        return {
-          ParameterKey: key,
-          ParameterValue: String(answers[key]),
-        };
-      }),
-      {
-        ParameterKey: 'Version',
-        ParameterValue: version,
-      },
-      ...(development
-        ? [
-            {
-              ParameterKey: 'ArtifactBucket',
-              ParameterValue: `wharfie-artifacts-${process.env.WHARFIE_REGION}`,
-            },
-          ]
-        : [
-            { ParameterKey: 'GitSha', ParameterValue: version },
-            {
-              ParameterKey: 'ArtifactBucket',
-              ParameterValue: `${process.env.WHARFIE_DEPLOYMENT_NAME}-${Account}-${process.env.WHARFIE_REGION}`,
-            },
-            {
-              ParameterKey: 'IsDevelopment',
-              ParameterValue: String(development),
-            },
-          ]),
-    ],
-    Capabilities: ['CAPABILITY_IAM'],
-    TemplateBody: JSON.stringify(template),
+  const deployment = new WharfieDeployment({
+    name: process.env.WHARFIE_DEPLOYMENT_NAME,
+    properties: answers,
   });
+  await deployment.reconcile();
   displaySuccess(`Wharfie deployment configuration successfully updated`);
 };
 
@@ -115,7 +62,6 @@ exports.handler = async function ({ development }) {
   try {
     await config(development);
   } catch (err) {
-    console.trace(err);
     displayFailure(err);
   }
 };
