--- cli/cmds/new_cmds/deployment.js
@@ -1,11 +1,0 @@
-'use strict';
-
-exports.command = 'deployment <create|upgrade|destroy|config>';
-exports.desc = 'manage the wharfie service deployment';
-exports.builder = function (yargs) {
-  return yargs
-    .commandDir('./deployment_cmds')
-    .demandCommand()
-    .showHelpOnFail(true);
-};
-exports.handler = function () {};
